# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_create_question_api_questions_post": {
        "properties": {
          "config_profile_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Profile Id"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_create_question_api_questions_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PaginatedQuestions": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/QuestionOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "skip": {
            "title": "Skip",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "skip",
          "limit"
        ],
        "title": "PaginatedQuestions",
        "type": "object"
      },
...
      "QuestionConfirmRequest": {
        "properties": {
          "confirmation_name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Confirmation Name",
            "type": "string"
          }
        },
        "required": [
          "confirmation_name"
        ],
        "title": "QuestionConfirmRequest",
        "type": "object"
      },
...
      "QuestionMutationResponse": {
        "properties": {
          "deleted_submission_count": {
            "title": "Deleted Submission Count",
            "type": "integer"
          }
        },
        "required": [
          "deleted_submission_count"
        ],
        "title": "QuestionMutationResponse",
        "type": "object"
      },
      "QuestionOut": {
        "properties": {
          "config_profile_id": {
            "title": "Config Profile Id",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "last_used_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "original_filename": {
            "title": "Original Filename",
            "type": "string"
          },
          "replacement_error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replacement Error Message"
          },
          "replacement_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QuestionReplacementStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/QuestionStatus"
          },
          "submission_count": {
            "default": 0,
            "title": "Submission Count",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "config_profile_id",
          "name",
          "original_filename",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "QuestionOut",
        "type": "object"
      },
...
      "QuestionReplacementStatus": {
        "enum": [
          "pending",
          "processing",
          "failed"
        ],
        "title": "QuestionReplacementStatus",
        "type": "string"
      },
      "QuestionStatus": {
        "enum": [
          "pending",
          "ocr_processing",
          "ready",
          "failed"
        ],
        "title": "QuestionStatus",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "SURF-2026-0031 后端 API",
    "title": "AI 作业批改系统",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/questions": {
      "get": {
        "operationId": "list_questions_api_questions_get",
        "parameters": [
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "default": "",
              "title": "Search",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/QuestionStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedQuestions"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Questions",
        "tags": [
          "questions"
        ]
      },
      "post": {
        "operationId": "create_question_api_questions_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_question_api_questions_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Question",
        "tags": [
          "questions"
        ]
      }
    },
    "/api/questions/{question_id}": {
      "delete": {
        "operationId": "delete_question_api_questions__question_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "title": "Question Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Question",
        "tags": [
          "questions"
        ]
      },
      "get": {
        "operationId": "get_question_api_questions__question_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "title": "Question Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Question",
        "tags": [
          "questions"
        ]
      },
      "patch": {
        "operationId": "rename_question_api_questions__question_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "title": "Question Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionRenameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rename Question",
        "tags": [
          "questions"
        ]
      }
    },
    "/api/questions/{question_id}/config-profile": {
      "patch": {
        "description": "切换题目使用的配置项目。",
        "operationId": "change_question_config_profile_api_questions__question_id__config_profile_patch",
        "parameters": [
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "title": "Question Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionConfigProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Change Question Config Profile",
        "tags": [
          "questions"
        ]
      }
    },
    "/api/questions/{question_id}/grading-prompt": {
      "get": {
        "description": "按题目生成可审计的批改提示词。\n\n复用与运行时评分包完全相同的 rubric 解析与评分策略组装\n（``resolve_submission_rubric`` + ``build_grading_policy``），\n保证教师复制的提示词与 ``open_ai_marking_assignment`` 返回的\ngrading_policy 同源、一字不差。题目 OCR 未完成时无法生成。",
        "operationId": "get_question_grading_prompt_api_questions__question_id__grading_prompt_get",
        "parameters": [
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "title": "Question Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GradingPromptOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Question Grading Prompt",
        "tags": [
          "questions"
        ]
      }
    },
    "/api/questions/{question_id}/pdf": {
      "get": {
        "operationId": "get_question_pdf_api_questions__question_id__pdf_get",
        "parameters": [
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "title": "Question Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Question Pdf",
        "tags": [
          "questions"
        ]
      }
    },
    "/api/questions/{question_id}/replace": {
      "post": {
        "operationId": "replace_question_api_questions__question_id__replace_post",
        "parameters": [
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "title": "Question Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_replace_question_api_questions__question_id__replace_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionReplacementResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replace Question",
        "tags": [
          "questions"
        ]
      }
    },
    "/api/questions/{question_id}/retry-ocr": {
      "post": {
        "description": "失败后由教师重新上传 PDF，再创建一次 OCR 任务。",
        "operationId": "retry_question_ocr_api_questions__question_id__retry_ocr_post",
        "parameters": [
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "title": "Question Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_retry_question_ocr_api_questions__question_id__retry_ocr_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Question Ocr",
        "tags": [
          "questions"
        ]
      }
    },
...
  }
}